+++ packages/sdk-coverage-tests/test/report.spec.js
@@ -104,7 +104,8 @@ describe('Report', () => {
     assert.deepStrictEqual(convertSdkNameToReportName('eyes-images'), 'js_images')
   })
   it('should convert xml report to QA report schema as JSON', () => {
-    assert.deepStrictEqual(convertJunitXmlToResultSchema({junit, metadata}), [
+    const result = convertJunitXmlToResultSchema({junit, metadata})
+    assert.deepStrictEqual(result, [
       {
         test_name: 'test check window with css',
         parameters: {
@@ -147,6 +148,17 @@ describe('Report', () => {
         },
         passed: undefined,
         isSkipped: true,
+        isGeneric: true,
+      },
+      {
+        test_name: 'test that was emitted but not executed',
+        parameters: {
+          browser: 'chrome',
+          mode: 'bla',
+          api: undefined,
+        },
+        passed: true,
+        isSkipped: false,
         isGeneric: true,
       },
     ])
@@ -200,6 +212,17 @@ describe('Report', () => {
           },
           passed: undefined,
           isSkipped: true,
+          isGeneric: true,
+        },
+        {
+          test_name: 'test that was emitted but not executed',
+          parameters: {
+            browser: 'chrome',
+            mode: 'bla',
+            api: undefined,
+          },
+          passed: true,
+          isSkipped: false,
           isGeneric: true,
         },
       ],
@@ -259,9 +282,93 @@ describe('Report', () => {
             isSkipped: true,
             isGeneric: true,
           },
+          {
+            test_name: 'test that was emitted but not executed',
+            parameters: {
+              browser: 'chrome',
+              mode: 'bla',
+              api: undefined,
+            },
+            passed: true,
+            isSkipped: false,
+            isGeneric: true,
+          },
         ],
         id: '111111',
       },
     )
+  })
+
+  it('should create a report with custom coverage tests', () => {
+    const junit = loadFixture('multiple-suites-with-custom-tests.xml')
+    const result = convertJunitXmlToResultSchema({junit, metadata})
+    assert.deepStrictEqual(result, [
+      {
+        test_name: 'test check window with css',
+        parameters: {
+          browser: 'chrome',
+          mode: 'css',
+          api: 'classic',
+        },
+        passed: true,
+        isSkipped: false,
+        isGeneric: true,
+      },
+      {
+        test_name: 'test check window with vg',
+        parameters: {
+          browser: 'chrome',
+          mode: 'visualgrid',
+          api: undefined,
+        },
+        passed: undefined,
+        isSkipped: true,
+        isGeneric: true,
+      },
+      {
+        test_name: 'test check window with scroll',
+        parameters: {
+          browser: 'chrome',
+          mode: 'scroll',
+          api: undefined,
+        },
+        passed: undefined,
+        isSkipped: true,
+        isGeneric: true,
+      },
+      {
+        test_name: 'test that was not emitted',
+        parameters: {
+          browser: 'chrome',
+          mode: 'bla',
+          api: undefined,
+        },
+        passed: undefined,
+        isSkipped: true,
+        isGeneric: true,
+      },
+      {
+        test_name: 'test that was emitted but not executed',
+        parameters: {
+          browser: 'chrome',
+          mode: 'bla',
+          api: undefined,
+        },
+        passed: true,
+        isSkipped: false,
+        isGeneric: true,
+      },
+      {
+        test_name: 'some custom test',
+        parameters: {
+          api: undefined,
+          browser: 'chrome',
+          mode: undefined,
+        },
+        passed: false,
+        isSkipped: false,
+        isGeneric: false,
+      },
+    ])
   })
 })